requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
apsw==3.36.0.post1
apturl==0.5.2
argon2-cffi==21.1.0
async-timeout==4.0.3
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
css-parser==1.0.7
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
docker-pycreds==0.4.0
duplicity==0.8.21
entrypoints==0.4
fasteners==0.14.1
feedparser==6.0.8
filelock==3.13.1
fonttools==4.29.1
frozenlist==1.4.0
fs==2.4.12
fsspec==2023.10.0
future==0.18.2
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.40
gyp==0.1
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
httplib2==0.20.2
hydra-core==1.3.2
idna==3.3
ifaddr==0.1.7
imageio==2.31.1
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.3.2
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightning-utilities==0.10.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.7.1
matplotlib-inline==0.1.3
mechanize==0.4.7
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
multidict==6.0.4
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==2.8.7
notebook==6.4.8
numpy==1.23.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
overrides==7.3.1
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
py==1.10.0
py7zr==0.11.3+dfsg
pycairo==1.20.1
pychm==0.8.6
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.11.0
pycups==2.0.1
pyemd==1.0.0
Pygments==2.11.2
PyGObject==3.42.1
PyGSP==0.5.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pythran==0.10.0
pytorch-lightning==2.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
rdkit==2023.3.2
regex==2021.11.10
reportlab==3.6.8
repoze.lru==0.7
requests==2.25.1
requests-toolbelt==0.9.1
Routes==2.5.1
scikit-learn==1.3.2
scipy==1.11.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
sentry-sdk==1.38.0
setproctitle==1.3.3
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.3.1
sympy==1.9
systemd-python==234
terminado==0.13.1
testpath==0.5.0
texttable==1.6.4
threadpoolctl==3.2.0
torch==2.1.1
torch-geometric==2.3.1
torch_scatter==2.1.2
torch_sparse==0.6.18
torchmetrics==0.11.4
tornado==6.1
tqdm==4.65.0
traitlets==5.1.1
triton==2.1.0
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.1.0
usb-creator==0.3.7
wadllib==1.3.6
wandb==0.15.4
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.6
widgetsnbextension==2.0.0
xdg==5
xkit==0.0.0
yarl==1.9.3
zeroconf==0.38.3
zipp==1.0.0
